make GodRaysComponent::nbSteps an int

+++ components/data/GodRaysComponent.hpp
@@ -5,7 +5,7 @@
 namespace kengine {
 	struct GodRaysComponent {
 		float scattering = .1f;
-		float nbSteps = 10.f;
+		int nbSteps = 10;
 		float defaultStepLength = 25.f;
 		float intensity = 2.5f;
 	};
